Update flight_model.cfg
Changed reference speed of the landing gear for VNAV to register appropriate speeds. this reference speed is to ensure that the AP does not shoot to 280 out the gate when on approach and gear is down. speed will be about 180 KTS which is the appropriate speed to hold when landing gear deployed.