Add files via upload
This is an update to the master branch work please install these files they are in beta but will help with the AP as well as the ILS of the 747. to help other users please make sure to uninstall my first mod. Also maker sure to press APP twice to capture G/S this is a hard coded issue in the game and I cannot change it.